Added installation section with instructions on how to install Cuarzo Engine using npm. Also included an example on how to import and use the engine in a JavaScript project.
Included a "Getting Started" section with a link to the "Getting Started" page in the Cuarzo Engine repository. Encouraged users to follow the provided steps to start developing games with the engine.

Provided congratulations on the npm publication of Cuarzo Engine and expressed confidence that the updated installation section will help users get started quickly.